backend/requirements.txt
fastapi
uvicorn
websockets
numpy
scikit-learn
psutil
onnxruntime
onnx
joblib
pandas
python-multipart
py-cpuinfo
skl2onnx
watchdog